Updated CICE_InitMod.F90 for cice driver to set all tracer values to c0 over land after initialization.  This supports bit-for-bit testing on different pe counts and different decompositions where land block elmination might be different and results in all land tracer values to be zero on history and restart files.
Updated the test scripts and added a new -td feature that allows test results to be compared
with each other.  Refactored some of the scripts to make them more flexible for the -td option.
Changed the test name convention so machine is the first item in the name.  This is to make
it easier to use the -td option.

Modified tests so now there are only smoke and restart tests.  Added new settings (run5day,
run10day, run1year, etc) to support different length smoke tests.  Refactored the test
scripts so there is a test_smoke.script and a test_restart.script now.  Moved the main
test script from under tests to the scripts directory and renamed it to cice.test.setup.csh.
Updated the README.test.

Updated the base_suite file to add new tests and make testnames conform to new convention.

Changed the restart test to be a 10 day run with restart at day 5.

Added other new options to support different batch wall times (short, medium long), support
modifications for grids (gx1, gx3), to turn on the bfbflag (bfbflagT).

Added some new error checks to make sure create.case aborts if there is a problem detected.

Modified create.case -h so it now automatically generates the list of -s options and the
list of -t tests as part of the documentation.

These results are bit-for-bit over the ocean and not bit-for-bit over land for some tracers.