Remove $(INCS) variable from build rules
The -I flag is used by Fortran compilers to locate files
included via the "#include" preprocessor directive or the "include" statement,
or modules referenced via the "use" statement. In our codebase we only use
modules, and all modules files (*.mod) are put in the ${ICE_OBJDIR} directory,
where the compilation takes place. So there is no need for all source directories
(contained in the variable $(INCS), prefixed with -I) to be added to the build
rules. This shortens the length of the displayed compilation lines.