Build system improvements (#354)
* Add cice.make script, list makdep as prerequisite

* cice.make can be used to easily call make with the different targets
  present in the Makefile (db_files, db_flags, clean), or to pass
  different flags to make
* makdep is listed as a prerequisite of the *.d files, such that
  we don't need to call make makdep separately in cice.build

* Update db_files and db_flags rules in Makefile

* Remove $(INCS) variable from build rules

The -I flag is used by Fortran compilers to locate files
included via the "#include" preprocessor directive or the "include" statement,
or modules referenced via the "use" statement. In our codebase we only use
modules, and all modules files (*.mod) are put in the ${ICE_OBJDIR} directory,
where the compilation takes place. So there is no need for all source directories
(contained in the variable $(INCS), prefixed with -I) to be added to the build
rules. This shortens the length of the displayed compilation lines.

* Update documentation

* add cice.build argument capabilities

* update documentation

* update documentation

* update documentation

* cleanup of documentation and implementation

* update implementation based on reviews

* update documentation

* update documentation

* update documentation

* update documentation

* update Makefile and documentation