parse_namelist.sh: don't use `sed -i` without a backup
Some `sed` implementations (notably, on macOS) do not handle using
the `-i` option without a backup file. As such, the following argument
('-e') is used as an extension for the backup file, resulting in a
file named ${filename}.check-e  which does not get
deleted and stays in the case directory.

Add a backup extension and remove the backup file at the end of the
script.