doc: correct 'Pstar' units
'Pstar' was added to the namelist in c9cfebd (add Pstar and Cstar to
namelist (#475), 2020-06-26), but the wrong units were added to the
documentation (N/m instead of N/m^2). Also, it already appeared in the
index with the same wrong units.

Fix that.

While at it, fix the RST syntax for `k2`.

Reported-by: Jean-François Lemieux <jean-francois.lemieux@canada.ca>