Add sw_redist option to CICE (#497)
This integrates the sw_redist changes from Icepack into CICE.

Add three new namelist, sw_redist, sw_frac, and sw_dtemp.

The alt03 and alt04 test cases have been updated so that ktherm=1 is bfb. The default settings are such that ktherm=2 is bfb. This is the other piece to addressing issue #485 and CICE-Consortium/Icepack#280.