doc: correct description of box2001 test
The documentation states that this test case sets 'coriolis' to zero,
but 'configuration/scripts/options/set_nml.box2001' sets 'coriolis' to 'constant'.
Fix the documentation to be in line with the code.

While at it, add the correct unit for the Coriolis parameter (in both places
where it appears in the documentation).

Closes CICE-Consortium/CICE#509

Reported-by: Jean-François Lemieux <jean-francois.lemieux@canada.ca>