Update prognostic calendar variables to myear, mmonth, mday, msec.
They were nyr, month, mday, sec.  The "m" in front means model and it
creates more unique names in the source code than year, month, day, sec
which can be hard to grep for because there are so many matches.

This is bit-for-bit on a full test suite on cheyenne.