cicecore: add support for staggered atmoshperic levels
In order to support atmospheric input data given at different levels for
winds and scalars, introduce a new array 'zlvs' in module ice_flux, and
pass it down to 'icepack_atmo_boundary'.

Initialize 'zlvs' to the same value as 'zlvl' (10 metres) so as not to
change the standalone model answers.