Compute seabed stress at CD-grid locations (#20)
* ice_dyn_shared: fix subname for 'seabed_stress_factor_LKD'

* ice_dyn_shared: add optional 'grid_location' argument to seabed_stress_factor_LKD

In subsequent commits we want to compute the seabed stress factor
'Tb' not only at the U point (NE corner) but also at the E (center of
E face) and N (center of N face) points.

In order to dispatch the computation in this subroutine to different
code paths depending on the grid location (U, N or E), add an optional
argument 'grid_location' that can be used to indicate at which point we
want the factor to be computed. Default it to the 'U' point for
backwards compatibility, such that the existing calls do not have to be
changed.

Note that we need an additional 'l_grid_location' variable to set the
default value, as the Fortran standard does not allow setting the value
of an optional argument if it is not present.

Note also that 'icellu' was incorrectly referencing 'icetmask'. Fix that
by generalizing the description.

* ice_grid: add 'grid_neighbor_{min,max}' functions

Add functions 'grid_neighbor_min' and 'grid_neighbor_max', used to
compute the minimum/maximum of neighboring T-point values of a field at
the U, E or N location.

* ice_dyn_shared: use 'grid_neighbor_{min,max}' in seabed_stress_factor_LKD

Use the 'grid_neighbor_{min,max}' subroutines introduced in the
preceding commit to compute the min and max of neighbor values of
hwater, aice and vice at 'grid_location'. This generalizes the
subroutine for the CD-grid. A subsequent commit will call it twice to
compute the seabed stress factor at the E and N locations.

* ice_dyn_shared: fix subname for 'seabed_stress_factor_prob'

* ice_dyn_shared: add optional CD-grid arguments to 'seabed_stress_factor_prob'

In contrast to 'seabed_stress_factor_LKD', in
'seabed_stress_factor_prob' the seabed stress factor is computed at the
T location and only transfered to the U location at the end of the
subroutine. So for efficiency it does not make sense to call that
subroutine twice for computing the factor at the E and N locations.

Instead, add optional arguments for 'TbE' and 'TbN', as well as the
corresponding indices paraphernalia, and compute the factor at the E and
N location at the end of the subroutine depending on 'grid_system' and
the presence of the optional arguments.

Use 'grid_neighbor_max' to abstract away the operation of finding the
maximum value of neighboring cells.

* ice_dyn_evp: compute seabed stress factor at CD-grid locations

Leverage the changes in the previous commits to compute the seabed
stress factor at the E and N locations if grid_system == 'CD'.

For seabed_stress_method == 'LKD', simply call seabed_stress_factor_LKD
twice, once for each of the E and N locations.

For seabed_stress_method == 'probabilistic', call
seabed_stress_factor_prob with the additional arguments for the E and N
locations.