Add CD-grid stress arrays (#28)
* ice_flux: add stress arrays at CD-grid locations (T,U)

* ice_dyn_shared: initialize CD-grid stress arrays

* infrastructure: add CD-grid stress arrays to restarts

* ice_dyn_shared: generalize 'principal_stress' arguments names

In a subsequent commit we will call 'principal_stress' with the CD-grid
stress arrays 'stress{p,m,12}T' to compute the principal stresses at the
tracer point when using the CD grid.

In that light, remove the '_1' suffix from the stress arguments since
they won't always be located at the NE corner anymore.

* ice_history: compute 'sig[12P]' at T-point for CD-grid

For the CD-grid, compute the principal stresses and the ice pressure at
the T point by passing the appropriate arrays to 'principal_stress'.

These three history variables are computed using the NE-corner values on
the B-grid, but this is not reflected in the description of the fields
in the history output.

Add the location at which the stresses are computed to the comment
argument in the call to 'define_hist_field', for both the B and CD grid.

* ice_dyn_evp: pass 'stress{p,m,12}[TU]' to dyn_prep2

For the CD-grid, the 'stress{p,m,12}_[1-4]' arrays are not used. Pass
the CD-grid location stress arrays 'stress{p,m,12}[TU]' to 'dyn_prep2',
which zero-initializes them anywhere icetmask is zero.

Also, add a TODO in the `if (grid_type) == 'tripole'` block, since it is
not yet clear how the halo updates should be done for the new stress
arrays on the tripole grid.