Update mct/cesm1 driver, fix ss_tlt implementation (#726)
* Update ss_tltx, ss_tlty averaging/usage in evp, eap, vp
Update mct/cesm1 driver for snow mods and other recent changes
Remove USE_NETCDF cpp in io_pio2/ice_restart.F90, not needed or correct

* Add tripoleOnly flag to ice_HaloUpdate2DR8.  This is needed in very
limited situations where we want to ensure bit-for-bit across the seam,
but do NOT want to update the halo in general.  In particular, this is
needed for UAREA during initialization.  UAREA is explicitly computed
in the halo to avoid missing data.  In some cases, uarea computed
independently along the seam in the halo points results in roundoff
level differences.  This will produce a small error in the simulation
but more important, breaks the symmetry across the seam.  In that case,
we just want to sync up the halo, but we do NOT want to update the
independent UAREA halo computations elsewhere because of land block
elimination and other situations where blocks have no neighbors.