cice.setup: remove 'suite.jobs' at start of 'suite.submit'
When running a test suite that includes 'bfbcomp' tests, the test
script ('cice.test') queries the file 'suite.jobs' in the suite
directory to get the job ID of the test against which to compare for
bit-for-bit-ness.

If re-running the suite, 'suite.jobs' is not removed, so that the new
job IDs are appended to the existing file. This leads to syntax errors
when 'cice.test' looks for the job ID to compare against since the
'grep' call returns several matches.

Remove 'suite.jobs' at the start of the 'suite.submit' script generated
by 'cice.setup' to avoid that.