ice_dyn_vp: use 'global_sum_prod' for bit-for-bit log reproducibility
In the previous commits we ensured bit-for-bit reproducibility of the
outputs when using the VP solver.

Some global norms computed during the nonlinear iteration still use the
same non-reproducible pattern of summing over blocks locally before
performing the reduction. However, these norms are used only to monitor
the convergence in the log file, as well as to exit the iteration when
the required convergence level is reached ('nlres_norm'). Only
'nlres_norm' could (in theory) influence the output, but it is unlikely
that a difference due to floating point errors would influence the 'if
(nlres_norm < tol_nl)' condition used to exist the nonlinear iteration.

Change these remaining cases to also use 'global_sum_prod' to compute
the global norm, leading to bit-for-bit log reproducibility.

Make sure to zero-initialize the arrays F[xy] corresponding to the
non-linear residual vector (Ax - b), once again because global_sum_prod
loops over the whole array, and only ice points are currently
initialized (in 'residual_vec') for these arrays.

In debug mode, this avoids NaNs being involved in some computations
during the 'global_sum_prod' calls used to compute the norm of the
residual just after it's computed. This was only triggered in certain
tests of the 'decomp_suite'.