doc: mention VP solver is only reproducible using 'bfbflag'
The previous commits made sure that the model outputs as well as the log
file output are bit-for-bit reproducible when using the VP solver by
refactoring the code to use the existing 'global_sum_prod' subrooutine.

Add a note in the documentation mentioning that 'bfbflag' is required to
get bit-for-bit reproducible results under different decompositions /
MPI count when using the VP solver.

While at it, correct a typo in cice_index.rst.